+++ packages/core-js/features/instance/replace-all.js
@@ -1,9 +1,3 @@
-var replaceAll = require('../string/virtual/replace-all');
+var parent = require('../../stable/instance/replace-all');
 
-var StringPrototype = String.prototype;
-
-module.exports = function (it) {
-  var own = it.replaceAll;
-  return typeof it === 'string' || it === StringPrototype
-    || (it instanceof String && own === StringPrototype.replaceAll) ? replaceAll : own;
-};
+module.exports = parent;